devide components to modules
