+++ pages/api/variable.js
@@ -1,5 +1,3 @@
-export const name = "Kyo";
-
 export const url = {
     urlBlog:"https://www.s-gakuenblog.com/",
     urlGit:"https://github.com/POD-azlamarhyu",
